+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 0.2
 ---
 
-  * Add [@key] attribute.
+  * Add `[@key]` attribute.
+  * Add support for `Yojson.Safe.json` values.
 
 0.1
 ---